# docs/openapi.json
{
  "components": {
    "schemas": {
      "AuditEventOut": {
        "properties": {
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Metadata only; never derived from a request body.",
            "title": "Detail"
          },
          "event": {
            "description": "e.g. login.ok, login.failed, vault.uploaded, vault.conflict.",
            "title": "Event",
            "type": "string"
          },
          "ip": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ip"
          },
          "user_agent": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Agent"
          }
        },
        "required": [
          "event",
          "ip",
          "user_agent",
          "detail",
          "created_at"
        ],
        "title": "AuditEventOut",
        "type": "object"
      },
      "ConflictDetail": {
        "properties": {
          "detail": {
            "default": "version conflict",
            "title": "Detail",
            "type": "string"
          },
          "server_version": {
            "description": "The version the client must merge against.",
            "title": "Server Version",
            "type": "integer"
          }
        },
        "required": [
          "server_version"
        ],
        "title": "ConflictDetail",
        "type": "object"
      },
      "ConflictOut": {
        "description": "The 409 body from PUT /vault. `detail` is an object here, not a string,\nso the client can re-sync without a second request.",
        "example": {
          "detail": {
            "detail": "version conflict",
            "server_version": 4
          }
        },
        "properties": {
          "detail": {
            "$ref": "#/components/schemas/ConflictDetail"
          }
        },
        "required": [
          "detail"
        ],
        "title": "ConflictOut",
        "type": "object"
      },
      "ErrorOut": {
        "description": "The shape of every non-409 error response.",
        "example": {
          "detail": "invalid credentials"
        },
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "string"
          }
        },
        "required": [
          "detail"
        ],
        "title": "ErrorOut",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "VaultDeleteOut": {
        "properties": {
          "deleted": {
            "default": true,
            "title": "Deleted",
            "type": "boolean"
          }
        },
        "title": "VaultDeleteOut",
        "type": "object"
      },
      "VaultMetaOut": {
        "properties": {
          "size_bytes": {
            "description": "Ciphertext length. Padm\u00e9-padded, so it does not track entry count.",
            "title": "Size Bytes",
            "type": "integer"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "ISO-8601 UTC, or null if never written.",
            "title": "Updated At"
          },
          "version": {
            "description": "Monotonic; 0 means nothing is stored yet.",
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "version",
          "size_bytes"
        ],
        "title": "VaultMetaOut",
        "type": "object"
      },
      "VaultOut": {
        "properties": {
          "blob": {
            "description": "Base64 ciphertext, opaque to this server.",
            "title": "Blob",
            "type": "string"
          },
          "size_bytes": {
            "description": "Ciphertext length. Padm\u00e9-padded, so it does not track entry count.",
            "title": "Size Bytes",
            "type": "integer"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "ISO-8601 UTC, or null if never written.",
            "title": "Updated At"
          },
          "version": {
            "description": "Monotonic; 0 means nothing is stored yet.",
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "version",
          "size_bytes",
          "blob"
        ],
        "title": "VaultOut",
        "type": "object"
      },
      "VaultPutIn": {
        "example": {
          "base_version": 0,
          "blob": "WktWMQABAgMEBQYHCAkKCwwNDg8QERITFBUWFxgZGhs="
        },
        "properties": {
          "base_version": {
            "description": "The version this edit was based on. 0 asserts that no vault is stored yet.",
            "minimum": 0.0,
            "title": "Base Version",
            "type": "integer"
          },
          "blob": {
            "description": "Base64 ciphertext, at most 16 MiB decoded.",
            "maxLength": 25165824,
            "title": "Blob",
            "type": "string"
          }
        },
        "required": [
          "blob",
          "base_version"
        ],
        "title": "VaultPutIn",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "contact": {
      "name": "zkvault",
      "url": "https://github.com/ahmedazizabbassi/pwd_manager"
    },
...
    "license": {
      "name": "MIT"
    },
    "summary": "Stores encrypted vault blobs. The server cannot decrypt them, cannot search them, and offers no password recovery.",
    "title": "zkvault blind sync",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/vault": {
      "delete": {
        "description": "Irreversible: there is no server-side copy of the key, so nothing here can\nbe recovered afterwards. Requires ?confirm=DELETE.",
        "operationId": "delete_vault_api_v1_vault_delete",
        "parameters": [
          {
            "description": "Must be the literal string \"DELETE\".",
            "in": "query",
            "name": "confirm",
            "required": false,
            "schema": {
              "default": "",
              "description": "Must be the literal string \"DELETE\".",
              "title": "Confirm",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VaultDeleteOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Missing ?confirm=DELETE."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Missing, expired or rejected credentials."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Rate limit exceeded. Honour the Retry-After header."
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Delete the stored vault (irreversible)",
        "tags": [
          "vault"
        ]
      },
      "get": {
        "description": "Responds with an `ETag` carrying the version, so a conditional client can skip unchanged bodies.",
        "operationId": "download_api_v1_vault_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VaultOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Missing, expired or rejected credentials."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "This account has never uploaded a vault."
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Rate limit exceeded. Honour the Retry-After header."
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Download the encrypted vault",
        "tags": [
          "vault"
        ]
      },
      "put": {
        "description": "Optimistic concurrency: the write is accepted only if the client's base\nversion still matches the stored one. Otherwise 409 with the server version,\nand the client re-downloads, merges locally, and retries.\n\nLast-writer-wins would be the easy alternative and would silently destroy an\nentry added on another device -- unacceptable for this data.",
        "operationId": "upload_api_v1_vault_put",
        "parameters": [
          {
            "description": "Optional belt and braces: the same number as `base_version`, quoted, e.g. `\"3\"`.",
            "in": "header",
            "name": "If-Match",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Optional belt and braces: the same number as `base_version`, quoted, e.g. `\"3\"`.",
              "title": "If-Match"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VaultPutIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VaultMetaOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Empty blob, malformed If-Match, or If-Match disagreeing with base_version."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Missing, expired or rejected credentials."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConflictOut"
                }
              }
            },
            "description": "base_version is stale. Re-download, merge locally, and retry."
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Blob exceeds ZKVAULT_MAX_BLOB_BYTES."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Rate limit exceeded. Honour the Retry-After header."
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Upload a new version",
        "tags": [
          "vault"
        ]
      }
    },
    "/api/v1/vault/audit": {
      "get": {
        "description": "A user's own security events -- logins, conflicts, MFA changes.\n\nVisible to the account holder because unexplained logins are exactly what a\nvictim needs to see early.",
        "operationId": "my_audit_log_api_v1_vault_audit_get",
        "parameters": [
          {
            "description": "Most recent events to return. Clamped to 1-200.",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "description": "Most recent events to return. Clamped to 1-200.",
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AuditEventOut"
                  },
                  "title": "Response My Audit Log Api V1 Vault Audit Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Missing, expired or rejected credentials."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Rate limit exceeded. Honour the Retry-After header."
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Your own security events",
        "tags": [
          "vault"
        ]
      }
    },
    "/api/v1/vault/meta": {
      "get": {
        "description": "Cheap version probe, so a client can skip downloading an unchanged vault.",
        "operationId": "vault_meta_api_v1_vault_meta_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VaultMetaOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Missing, expired or rejected credentials."
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Rate limit exceeded. Honour the Retry-After header."
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Current version and size",
        "tags": [
          "vault"
        ]
      }
    },
...
  },
  "servers": [
    {
      "description": "Local dev server (make serve)",
      "url": "http://127.0.0.1:8000"
    }
  ],
...
}